+++ src/e2e/app/repo-tests/src/test_flightsql.rs
@@ -6,11 +6,3 @@
 // As of the Change Date specified in that file, in accordance with
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
-
-mod access_token_notifier;
-mod account_lifecycle_notifier;
-mod flow_progress_notifier;
-
-pub use access_token_notifier::*;
-pub use account_lifecycle_notifier::*;
-pub use flow_progress_notifier::*;